api/requirements.txt
aniso8601==6.0.0
astroid==2.5.6
autopep8==2.0.0
Click==7.0
disposable-email-domains==0.0.65
Flask==1.0.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
gunicorn==20.1.0
isort==5.9.1
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.8.3
python-dotenv==0.18.0
pytest==7.2.0
pytest-flask==1.2.0
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.6
toml==0.10.2
validate-email==1.3
validate-email-wt==1.4.6
Werkzeug==2.0.1
wrapt==1.12.1